Update to zig 0.13.0. Refactor highlight config parser.

+++ Makefile
@@ -3,7 +3,7 @@ ZIGFLAGS?=
 
 # zig does not automatically clean up the cache, so we should make
 # one in the temp dir
-ZIG_CACHE=zig-cache
+ZIG_CACHE=.zig-cache
 ZIG_CACHE_TMP=/tmp/zig-cache-zenith
 
 .PHONY: build
@@ -15,11 +15,11 @@ $(ZIG_CACHE):
 	@if [ -d "$(ZIG_CACHE)" ]; then \
 		(du -sb $(ZIG_CACHE_TMP) | awk '{size=$$1; if (size > 1000000000) { exit 1 }}'); \
 		if [ $$? -ne 0 ]; then \
-			echo "$(ZIG_CACHE) is too big! You should delete it"; \
+			echo "$(ZIG_CACHE) is too big! You should clean it up with make clean-cache"; \
 		fi \
 	else \
 		mkdir -p $(ZIG_CACHE_TMP); \
-		ln -sf $(ZIG_CACHE_TMP) $(ZIG_CACHE); \
+		[ ! -L "$(ZIG_CACHE)" ] && ln -sf $(ZIG_CACHE_TMP) $(ZIG_CACHE); \
 	fi
 
 .PHONY: release
